requirements.txt
about-time==3.1.1
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
alive-progress==2.4.1
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
cachetools==5.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.4
cliff==4.1.0
cloudpickle==2.2.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.6
docker-pycreds==0.4.0
filelock==3.8.2
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grapheme==0.6.0
greenlet==2.0.1
grpcio==1.51.1
hyperopt==0.2.7
idna==2.8
importlib-metadata==4.13.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lightning-utilities==0.3.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
msgpack==1.0.4
multidict==6.0.3
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
optuna==3.0.4
packaging==22.0
pandas==1.5.2
pathtools==0.1.2
pbr==5.11.0
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.11.0
prettytable==3.5.0
promise==2.3
protobuf==3.20.1
psutil==5.9.4
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.2
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.6
PyYAML==6.0
ray>=2.1.0
rdflib==6.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.8.1
seaborn==0.12.1
sentry-sdk==1.9.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.45
stevedore==4.1.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.1.1
threadpoolctl==3.1.0
torchmetrics==0.11.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.25.11
virtualenv==20.17.1
wandb==0.13.6
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
